\name{nicepackage-package}
\alias{nicepackage-package}
\alias{nicepackage}
\docType{package}
\title{
\packageTitle{nicepackage}
}
\description{
\packageDescription{nicepackage}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{nicepackage}

Maintainer: \packageMaintainer{nicepackage}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = nicepackage)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{nicepackage}}
%\section{Documentation Index}{\packageIndices{nicepackage}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
